Refactor Engine and Browser to separate concerns
